requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
